+++ front/app/modules/commercial/content_builder/admin/components/ContentBuilderTopBar/index.tsx
@@ -45,8 +45,8 @@ import { Locale } from 'typings';
 type ContentBuilderTopBarProps = {
   hasPendingState?: boolean;
   localesWithError: Locale[];
-  mobilePreviewEnabled: boolean;
-  setMobilePreviewEnabled: React.Dispatch<React.SetStateAction<boolean>>;
+  previewEnabled: boolean;
+  setPreviewEnabled: React.Dispatch<React.SetStateAction<boolean>>;
   selectedLocale: Locale | undefined;
   draftEditorData?: Record<string, SerializedNodes>;
   onSelectLocale: (args: {
@@ -57,8 +57,8 @@ type ContentBuilderTopBarProps = {
 
 const ContentBuilderTopBar = ({
   params: { projectId },
-  mobilePreviewEnabled,
-  setMobilePreviewEnabled,
+  previewEnabled,
+  setPreviewEnabled,
   selectedLocale,
   onSelectLocale,
   draftEditorData,
@@ -105,6 +105,11 @@ const ContentBuilderTopBar = ({
       }
       setLoading(false);
     }
+  };
+
+  const redirectToProject = async () => {
+    await handleSave();
+    window.open(`/projects/${project?.attributes.slug}`, '_blank');
   };
 
   const handleSelectLocale = (locale: Locale) => {
@@ -166,11 +171,21 @@ const ContentBuilderTopBar = ({
         )}
         <Box ml="24px" />
         <Toggle
-          id="e2e-mobile-preview-toggle"
-          label={<FormattedMessage {...messages.mobilePreview} />}
-          checked={mobilePreviewEnabled}
-          onChange={() => setMobilePreviewEnabled(!mobilePreviewEnabled)}
+          id="e2e-preview-toggle"
+          label={<FormattedMessage {...messages.preview} />}
+          checked={previewEnabled}
+          onChange={() => setPreviewEnabled(!previewEnabled)}
         />
+        <Button
+          id="e2e-view-project-button"
+          buttonStyle="secondary"
+          icon="eye"
+          mx="20px"
+          disabled={!project}
+          onClick={redirectToProject}
+        >
+          <FormattedMessage {...messages.viewProject} />
+        </Button>
         <Button
           disabled={disableSave || hasPendingState}
           id="e2e-content-builder-topbar-save"
@@ -178,7 +193,6 @@ const ContentBuilderTopBar = ({
           processing={loading}
           onClick={handleSave}
           data-testid="contentBuilderTopBarSaveButton"
-          ml="24px"
         >
           <FormattedMessage {...messages.contentBuilderSave} />
         </Button>