+++ front/app/modules/commercial/content_builder/admin/containers/index.tsx
@@ -11,7 +11,7 @@ import { stylingConsts, colors } from 'utils/styleUtils';
 import { RightColumn } from 'containers/Admin';
 import { Box } from '@citizenlab/cl2-component-library';
 import Error from 'components/UI/Error';
-import ContentBuilderMobileView from '../components/ContentBuilderMobileView';
+import ContentBuilderEditModePreview from '../components/ContentBuilderEditModePreview';
 
 // craft
 import Editor from '../components/Editor';
@@ -58,7 +58,7 @@ export const CONTENT_BUILDER_DELETE_ELEMENT_EVENT =
   'deleteContentBuilderElement';
 
 export const ContentBuilderPage = () => {
-  const [mobilePreviewEnabled, setMobilePreviewEnabled] = useState(false);
+  const [previewEnabled, setPreviewEnabled] = useState(false);
   const [selectedLocale, setSelectedLocale] = useState<Locale | undefined>();
   const [draftData, setDraftData] = useState<Record<string, SerializedNodes>>();
   const { pathname } = useLocation();
@@ -197,15 +197,15 @@ export const ContentBuilderPage = () => {
           <ContentBuilderTopBar
             localesWithError={localesWithError}
             hasPendingState={imageUploading}
-            mobilePreviewEnabled={mobilePreviewEnabled}
-            setMobilePreviewEnabled={setMobilePreviewEnabled}
+            previewEnabled={previewEnabled}
+            setPreviewEnabled={setPreviewEnabled}
             selectedLocale={selectedLocale}
             onSelectLocale={handleSelectedLocaleChange}
             draftEditorData={draftData}
           />
           <Box
             mt={`${stylingConsts.menuHeight}px`}
-            display={mobilePreviewEnabled ? 'none' : 'flex'}
+            display={previewEnabled ? 'none' : 'flex'}
           >
             {selectedLocale && (
               <ContentBuilderToolbox selectedLocale={selectedLocale} />
@@ -230,11 +230,11 @@ export const ContentBuilderPage = () => {
             <ContentBuilderSettings />
           </Box>
         </Editor>
-        <Box
-          justifyContent="center"
-          display={mobilePreviewEnabled ? 'flex' : 'none'}
-        >
-          <ContentBuilderMobileView projectId={projectId} ref={iframeRef} />
+        <Box justifyContent="center" display={previewEnabled ? 'flex' : 'none'}>
+          <ContentBuilderEditModePreview
+            projectId={projectId}
+            ref={iframeRef}
+          />
         </Box>
       </FocusOn>
     </Box>